+++ app/code/community/Vaimo/Maksuturva/Model/Discount.php
@@ -1,193 +1,112 @@
 <?php
 /**
- * Copyright (c) 2016 Vaimo Finland Oy, Suomen Maksuturva Oy
+ * Copyright (c) 2018 Vaimo Finland Oy, Suomen Maksuturva Oy
  * See LICENSE.txt for license details.
  */
 
-class Vaimo_Maksuturva_Model_Discount extends Mage_SalesRule_Model_Quote_Discount
+class Vaimo_Maksuturva_Model_Discount extends Mage_SalesRule_Model_Validator
 {
-    private $configs;
+    protected $methodDiscounts;
 
     /**
-     * Collect address discount amount
+     * Returns array with amount and type if discount exists, false otherwise
      *
-     * @param   Mage_Sales_Model_Quote_Address $address
-     * @return  Mage_SalesRule_Model_Quote_Discount
+     * @param $methodCode
+     * @return array|bool
      */
-    public function collect(Mage_Sales_Model_Quote_Address $address)
+    public function getMethodDiscount($methodCode)
     {
-        parent::collect($address);
-
-        $active = $this->getConfigs()['active'];
-
-        if (true == $active) {
-            $items = $this->_getAddressItems($address);
-            $this->processMaksuturvaDiscount($address, $items);
-
-            $this->_calculator->prepareDescription($address);
+        if ($this->methodDiscounts == null) {
+            $this->methodDiscounts = $this->getMethodDiscounts();
         }
 
-        return $this;
+        if (!isset($this->methodDiscounts[$methodCode])) {
+            return false;
+        }
+        return $this->methodDiscounts[$methodCode];
+    }
+
+    /**
+     * @return array
+     */
+    public function getMethodDiscounts()
+    {
+        $address = Mage::getSingleton('checkout/cart')->getQuote()->getShippingAddress();
+
+        $discounts = [];
+
+        if (!$address instanceof Mage_Sales_Model_Quote_Address) {
+            return $discounts;
+        }
+
+        $store = Mage::app()->getStore($address->getStoreId());
+
+        $this->init($store->getWebsiteId(), $address->getCustomerGroupId(), $address->getCouponCode());
+
+        return $this->getAddressMethodDiscounts($address);
     }
 
     /**
      * @param $address Mage_Sales_Model_Quote_Address
-     * @param $items
+     * @return array
      */
-    protected function processMaksuturvaDiscount($address, $items)
+    protected function getAddressMethodDiscounts($address)
     {
-        $discount = $this->getPaymentDiscount($address->getQuote()->getPayment());
-        $discountPercent = $discount['amount'];
-        $discountDescription = $discount['description'];
+        // Payment method will be changed to check which discounts pass with Maksuturva methods
+        // Cloning address to avoid changing payment method of original address
+        $address = clone $address;
 
-        if ($discountPercent <= 0 || empty($discountDescription)) {
-            return;
-        }
+        $methodDiscounts = [];
 
-        foreach ($items as $item) {
-            $itemPrice = $item->getDiscountCalculationPrice();
-            $baseItemPrice = $item->getDiscountCalculationPrice();
-            $qty = $item->getQty();
+        foreach ($this->_getRules() as $rule) {
+            $methodCode = $this->getMaksuturvaPaymentMethodCode($rule);
 
-            $discountAmount = ($qty * $itemPrice - $item->getDiscountAmount()) * $discountPercent / 100;
-            $baseDiscountAmount  = ($qty * $baseItemPrice - $item->getBaseDiscountAmount()) * $discountPercent / 100;
-
-            $item->setDiscountAmount($discountAmount);
-            $item->setBaseDiscountAmount($baseDiscountAmount);
-
-            $this->_aggregateItemDiscount($item);
-        }
-
-        $shippingDiscount = ($address->getShippingAmount() - $address->getShippingDiscountAmount()) * $discountPercent / 100;
-        $baseShippingDiscount = ($address->getBaseShippingAmount() - $address->getBaseShippingDiscountAmount()) * $discountPercent / 100;
-
-        $this->_addAmount(-$shippingDiscount);
-        $this->_addBaseAmount(-$baseShippingDiscount);
-
-        if ($address->getDiscountDescription()) {
-            $address->setDiscountDescription($address->getDiscountDescription() . ','. $discountDescription);
-        } else {
-            $address->setDiscountDescription($discountDescription);
-        }
-    }
-
-    /**
-     * @return array Example: ['FI01' => '5.0', 'FI55' => '10']
-     */
-    public function getDiscounts()
-    {
-        $discountConfigStr = $this->getConfigs()['methodDiscounts'];
-        $discountConfigStr = preg_replace('/\s+/', '', $discountConfigStr);
-        $discounts = explode(',', $discountConfigStr);
-
-        $discountArr = [];
-        foreach ($discounts as $discount) {
-            $discountTuple = explode('=', $discount);
-
-            if (count($discountTuple) != 2) {
+            if (false == $methodCode) {
                 continue;
             }
-            $discountArr[$discountTuple[0]] = $discountTuple[1];
-        }
-        return $discountArr;
-    }
 
-    /**
-     * @param $methodCode
-     *
-     * @return float
-     */
-    public function getMethodDiscount($methodCode)
-    {
-        $discounts = $this->getDiscounts();
+            // Set Maksuturva method so rule passes if all other conditions are met
+            $this->setMaksuturvaPaymentMethod($address, $methodCode);
 
-        if (isset($discounts[$methodCode])) {
-            return $discounts[$methodCode];
-        }
-        return 0.0;
-    }
-
-    /**
-     * @param Mage_Sales_Model_Quote_Payment $payment
-     *
-     * @return array|false
-     */
-    public function getPaymentDiscount($payment)
-    {
-        $preselectMethod = $this->getPreselectMethod($payment);
-
-        if (! $preselectMethod) {
-            return false;
-        }
-
-        $amount = $this->getMethodDiscount($preselectMethod);
-
-        if ($amount <= 0) {
-            return false;
-        }
-
-        $description = $this->getDiscountDescription($preselectMethod, $amount);
-
-        return array(
-            'amount' => $amount,
-            'description' => $description
-        );
-    }
-
-    /**
-     * @param string $methodCode
-     * @param float $amount
-     *
-     * @return string
-     */
-    protected function getDiscountDescription($methodCode, $amount)
-    {
-        $methodName = $this->getPaymentMethodName($methodCode);
-
-        // PM - Payment Method
-        return sprintf('%s-%.01f%%', $methodName, $amount);
-    }
-
-    /**
-     * @param Mage_Sales_Model_Quote_Payment $payment
-     *
-     * @return mixed
-     */
-    protected function getPreselectMethod($payment)
-    {
-        if ($this->getConfigs()['preselectPaymentMethod']) {
-            $additional_data = unserialize($payment->getAdditionalData());
-            if (isset($additional_data[Vaimo_Maksuturva_Model_Maksuturva::MAKSUTURVA_PRESELECTED_PAYMENT_METHOD])) {
-                return $additional_data[Vaimo_Maksuturva_Model_Maksuturva::MAKSUTURVA_PRESELECTED_PAYMENT_METHOD];
+            if ($this->_canProcessRule($rule, $address)) {
+                if (!isset($methodDiscounts[$methodCode])) {
+                    $methodDiscounts[$methodCode] = [
+                        'amount' => $rule->getDiscountAmount(),
+                        'type' => $rule->getSimpleAction()
+                    ];
+                }
             }
         }
-
-        return false;
+        return $methodDiscounts;
     }
 
-    protected function getPaymentMethodName($methodCode)
+    /**
+     * @param $address Mage_Sales_Model_Quote_Address
+     * @param $method string
+     */
+    protected function setMaksuturvaPaymentMethod($address, $methodCode)
     {
-        $maksuturva = Mage::getModel('maksuturva/maksuturva');
-        $methods = $maksuturva->getPaymentMethods();
+        $payment = $address->getQuote()->getPayment();
+        $additional_data = unserialize($payment->getAdditionalData());
+        $additional_data[Vaimo_Maksuturva_Model_Maksuturva::MAKSUTURVA_PRESELECTED_PAYMENT_METHOD] = $methodCode;
+        $payment->setAdditionalData(serialize($additional_data));
+    }
 
-        foreach ($methods as $method) {
-            if ($method->{'code'} == $methodCode) {
-                return $method->{'displayname'};
+    /**
+     * @param $rule Mage_SalesRule_Model_Rule
+     * @return bool
+     */
+    protected function getMaksuturvaPaymentMethodCode($rule)
+    {
+        $methodCode = false;
+
+        $conditions = $rule->getConditions();
+
+        foreach ($conditions->getConditions() as $condition) {
+            if ($condition->getAttribute() == 'maksuturva_payment_method') {
+                $methodCode = $condition->getValue();
             }
         }
         return $methodCode;
-    }
-
-    protected function getConfigs()
-    {
-        if (is_null($this->configs)) {
-            $maksuturva = Mage::getModel('maksuturva/maksuturva');
-            $this->configs = [
-                'preselectPaymentMethod' => $maksuturva->getConfigData('preselect_payment_method'),
-                'methodDiscounts' => $maksuturva->getConfigData('method_discounts'),
-                'active' => $maksuturva->getConfigData('active')
-            ];
-        }
-        return $this->configs;
     }
 }